Saving Data for reloading Git
