+++ app/src/main/res/layout/login_page.xml
@@ -1,0 +1,114 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@drawable/dalhousie"
+    >
+
+    <android.support.constraint.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="#AAFFFFFF"
+        app:layout_constraintBottom_toTopOf="parent"
+        app:layout_constraintTop_toBottomOf="parent"
+        app:layout_constraintVertical_bias="0.3">
+
+
+        <TextView
+            android:id="@+id/promptID"
+            android:layout_width="100dp"
+            android:layout_height="50dp"
+            android:background="@color/brightGrey"
+            android:gravity="center"
+            android:text="@string/ids"
+            android:textStyle="bold"
+            app:layout_constraintEnd_toStartOf="@+id/idInput"
+            app:layout_constraintHorizontal_bias="0.5"
+            app:layout_constraintStart_toStartOf="parent" />
+
+        <EditText
+            android:id="@+id/idInput"
+            android:layout_width="250dp"
+            android:layout_height="50dp"
+            android:layout_marginStart="5dp"
+            android:hint="@string/idHint"
+            android:inputType="text"
+            android:textStyle="bold"
+            android:textSize="15sp"
+            app:layout_constraintBottom_toBottomOf="@id/promptID"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.5"
+            app:layout_constraintStart_toEndOf="@+id/promptID" />
+
+        <TextView
+            android:id="@+id/promptPW"
+            android:layout_width="100dp"
+            android:layout_height="50dp"
+            android:layout_marginTop="5dp"
+            android:background="@color/brightGrey"
+            android:textStyle="bold"
+            android:gravity="center"
+            android:text="@string/password"
+            app:layout_constraintEnd_toStartOf="@+id/pwInput"
+            app:layout_constraintHorizontal_bias="0.5"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/promptID" />
+
+        <EditText
+            android:id="@+id/pwInput"
+            android:layout_width="250dp"
+            android:layout_height="50dp"
+            android:layout_marginStart="5dp"
+            android:hint="@string/password"
+            android:inputType="textPassword"
+            android:textStyle="bold"
+            android:textSize="15sp"
+            app:layout_constraintBottom_toBottomOf="@+id/promptPW"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.5"
+            app:layout_constraintStart_toEndOf="@+id/promptPW" />
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_below="@+id/submitBT"
+            android:layout_centerHorizontal="true"
+            android:text="@string/signinMSG"
+            android:textStyle="bold"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.5"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/promptPW" />
+
+        <Button
+            android:id="@+id/submitBT"
+            android:layout_width="100dp"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="20dp"
+            android:text="@string/submitBT"
+            android:textStyle="bold"
+            app:layout_constraintEnd_toStartOf="@+id/cancelBT"
+            app:layout_constraintHorizontal_bias="0.5"
+            app:layout_constraintHorizontal_chainStyle="packed"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/promptPW" />
+
+        <Button
+            android:id="@+id/cancelBT"
+            android:layout_width="100dp"
+            android:layout_height="wrap_content"
+            android:layout_below="@+id/promptPW"
+            android:layout_centerHorizontal="true"
+            android:layout_marginTop="20dp"
+            android:text="@string/cancelBT"
+            android:textStyle="bold"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.5"
+            app:layout_constraintStart_toEndOf="@+id/submitBT"
+            app:layout_constraintTop_toBottomOf="@id/promptPW" />
+
+    </android.support.constraint.ConstraintLayout>
+
+</android.support.constraint.ConstraintLayout>
